Moved AIC, BIC, and LME functions to a new metrics module.
