+++ fitr/fitr.py
@@ -42,7 +42,8 @@ from scipy.optimize import minimize
 from scipy.optimize import brute
 from scipy.stats import multivariate_normal as mvn
 
-from .utils import trans_UC, BIC, AIC, LME
+from .utils import trans_UC
+from .metrics import BIC, AIC, LME
 
 # ==============================================================================
 #
