+++ fitr/utils.py
@@ -132,60 +132,3 @@ def trans_UC(values_U, rng):
         elif rng == 'unc': # Range: - inf - inf.
             values_T.append(value)
     return np.array(values_T)
-
-def BIC(loglik, nparams, nsteps):
-    """
-    Calculates Bayesian information criterion
-
-    Parameters
-    ----------
-    loglik : float or ndarray(dtype=float)
-        Log-likelihood
-    nparams : int
-        Number of parameters in the model
-    nsteps : int
-        Number of time steps in the task
-
-    Returns
-    -------
-    float or ndarray(dtype=float)
-
-    """
-    return nparams*np.log(nsteps) - 2*loglik
-
-def AIC(nparams, loglik):
-    """
-    Calculates Aikake information criterion
-
-    Parameters
-    ----------
-    nparams : int
-        Number of parameters in the model
-    loglik : float or ndarray(dtype=float)
-        Log-likelihood
-
-    Returns
-    -------
-    float or ndarray(dtype=float)
-
-    """
-    return 2*nparams - 2*loglik
-
-def LME(logpost, nparams, hessian):
-    """
-    Calculates log-model-evidence (LME)
-
-    Parameters
-    ----------
-    logpost : float or ndarray(dtype=float)
-        Log-posterior probability
-    nparams : int
-        Number of parameters in the model
-    hessian : ndarray(size=(nparams, nparams))
-        Hessian computed from parameter optimization
-
-    Returns
-    -------
-    float or ndarray(dtype=float)
-    """
-    return logpost + (nparams/2)*np.log(2*np.pi)-np.log(np.linalg.det(hessian))/2
